extract <PrevNextNavigation> - shared between blog-post & contentful-blog-post templates - pass `gradient` as props.nextGradient which is used to animate the next button - update PREV/NEXT styling with MUI-like box shadow - add typescript Doc Comments & interface definitions
