+++ src/components/TemplateComponents/index.tsx
@@ -1,2 +1,3 @@
 export { default as Discussion } from "./Discussion"
 export { default as DocumentToReactComponents } from "./Contentful"
+export { default as PrevNextNavigation } from "./PrevNextNavigation"
